Create actions folder and reorganize files accordingly
