+++ examples/medplum-scheduling-demo/src/pages/AppointmentDetailPage.tsx
@@ -3,7 +3,7 @@ import { Appointment, Patient, Reference } from '@medplum/fhirtypes';
 import { Document, PatientSummary, useResource } from '@medplum/react';
 import { useParams } from 'react-router-dom';
 import { AppointmentDetails } from '../components/AppointmentDetails';
-import { AppointmentActions } from '../components/AppointmentActions';
+import { AppointmentActions } from '../components/actions/AppointmentActions';
 
 export function AppointmentDetailPage(): JSX.Element {
   const { id } = useParams();
