+++ examples/medplum-scheduling-demo/src/pages/PatientPage.tsx
@@ -2,7 +2,7 @@ import { Grid, Loader } from '@mantine/core';
 import { Document, PatientSummary, useResource } from '@medplum/react';
 import { useParams } from 'react-router-dom';
 import { PatientDetails } from '../components/PatientDetails';
-import { PatientActions } from '../components/PatientActions';
+import { PatientActions } from '../components/actions/PatientActions';
 import { Patient } from '@medplum/fhirtypes';
 
 export function PatientPage(): JSX.Element {
